pyproject.toml
[tool.poetry]
name = "django-template"
version = "0.1.0"
description = "This template is build to accelerate development in dev and migration to stage and production."
authors = ["Mohd Zakir Hussain <mohd.zakir.hussain@outlook.com>"]
readme = "README.md"
packages = [
    { include = "apps" },
    { include = "config" }
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
django = ">=5.2.1,<6.0.0"
djangorestframework = ">=3.16.0,<4.0.0"
django-environ = ">=0.12.0,<0.13.0"
drf-spectacular = ">=0.28.0,<0.29.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

